fix: split stores into domain-specific modules
- Split store/index.js into separate domain modules: auth, theme, ui, chat, course, group, video, notification
- Fix: addToWishlist now guards against duplicates
- Fix: normalizeTheme handles corrupted/legacy persisted data
- Fix: useThemeStore toggles work with invalid stored data
- Build verified
